opt: fix tests for 64-bit systems.
Actually, only an issue for 64 bit big endian systems, but still...

+++ ccan/opt/test/run-consume_words.c
@@ -7,7 +7,7 @@
 /* Test consume_words helper. */
 int main(int argc, char *argv[])
 {
-	unsigned int start, len;
+	size_t start, len;
 
 	plan_tests(13);
